+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 
 ### Deprecations
 
-* `netket.vqs.ExactState` has been renamed to {class}`netket.vqs.FullSummationState` to better reflect what it does. Using the old name will now raise a warning [#14XX](https://github.com/netket/netket/pull/14XX).
+* `netket.vqs.ExactState` has been renamed to {class}`netket.vqs.FullSumState` to better reflect what it does. Using the old name will now raise a warning [#14XX](https://github.com/netket/netket/pull/14XX).
 
 
 ## NetKet 3.8 (⚙️ 8 May 2023)